Refactor to more beatyful code. Minor improvement
