+++ frontend/App.h
@@ -1,0 +1,10 @@
+void App();
+
+void Panel(int y0, int x0, int height, int width);
+void Chat(int y0, int x0, int height, int width);
+void ChatList(int y0, int x0, int height, int width);
+void JoinButton(int y0, int x0, int height, int width);
+void CreateChat(int y0, int x0, int height, int width);
+void ChatHeader(int y0, int x0, int height, int width);
+void ChatBlock(int y0, int x0, int height, int width);
+void InputField(int y0, int x0, int height, int width);
